vpu5: Eliminate compile warnings
Mostly data type conversions and void pointer arithmetic warnings with the
following exceptions:

* shvpu5_avcdec_notify.c and shvpu5_avcdec_omx.c - ambiguous else condition
* shvpu5_avcdec_omx.c - definition of ref_count lacked type (set to int)
			split switch statements for OMX_REVPU5INDEXTYPE cases
* shvpu5_avcenc_encode.c - avcenv_set_Q_matric second argument made non-const
* shvpu5_common_driver.c - handle_shvpu5_interrupt return type changed to
                           void *

+++ src/vpu5/shvpu5_avcdec_output.c
@@ -49,7 +49,7 @@ mcvdec_uf_get_frame_memory(MCVDEC_CONTEXT_T *context,
 	long fmem_x;
 	int i;
 	void *ypic_vaddr;
-	unsigned int ypic_paddr, cpic_paddr;
+	unsigned long ypic_paddr, cpic_paddr;
 	unsigned long align, alloc_size;
         shvpu_avcdec_PrivateType *shvpu_avcdec_Private =
                 (shvpu_avcdec_PrivateType *)context->user_info;
